requirements.txt
flask>=2.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
lxml>=4.6.0
python-dotenv>=0.19.0
gunicorn>=20.1.0
